requirements.txt
asgiref==3.6.0
click==8.1.3
colorama==0.4.6
Django==4.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
flake8==6.0.0
joblib==1.2.0
mccabe==0.7.0
nltk==3.8.1
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
pytz==2023.3
regex==2023.3.23
sqlparse==0.4.4
tqdm==4.65.0
tzdata==2023.3
